pyproject.toml
[tool.poetry]
name = "larin-coursework-6-skymarket"
version = "0.1.0"
description = "coursework6"
authors = ["Larin"]
readme = "README.md"
packages = [{include = "larin_coursework_6_skymarket"}]

[tool.poetry.dependencies]
python = "^3.11"
req2toml = "^1.2.0"
asgiref = "^3.6.0"
certifi = "^2022.12.7"
cffi = "^1.15.1"
charset-normalizer = "^3.0.1"
coreapi = "^2.3.3"
coreschema = "^0.0.4"
cryptography = "^38.0.4"
defusedxml = "^0.7.1"
django = "^4.1.4"
django-cors-headers = "^3.13.0"
django-filter = "^22.1"
django-phonenumber-field = "^7.0.1"
django-templated-mail = "^1.1.1"
djangorestframework = "^3.14.0"
idna = "^3.4"
inflection = "^0.5.1"
isort = "^5.11.4"
itypes = "^1.2.0"
jinja2 = "^3.1.2"
markupsafe = "^2.1.1"
mypy-extensions = "^0.4.3"
oauthlib = "^3.2.2"
packaging = "^22.0"
pathspec = "^0.10.3"
phonenumbers = "^8.13.3"
pillow = "^9.3.0"
platformdirs = "^2.6.0"
psycopg2 = "^2.9.5"
pycparser = "^2.21"
pyjwt = "^2.6.0"
pyparsing = "^3.0.9"
python-dotenv = "^0.21.0"
python3-openid = "^3.2.0"
pytz = "^2022.7"
six = "^1.16.0"
social-auth-app-django = "^5.0.0"
social-auth-core = "^4.3.0"
sqlparse = "^0.4.3"
tomli = "^2.0.1"
uritemplate = "^4.1.1"
drf-spectacular = "^0.25.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
